lab-04-secrets-sops/app/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pyyaml==6.0.1
